+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-content-client-v1.8.20](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.8.19...@adobe/spacecat-shared-content-client-v1.8.20) (2026-03-28)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1477](https://github.com/adobe/spacecat-shared/issues/1477)) ([67bdd1a](https://github.com/adobe/spacecat-shared/commit/67bdd1a2c497bed088bc1e54ae22e60c171308d1))
+
 ## [@adobe/spacecat-shared-content-client-v1.8.19](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.8.18...@adobe/spacecat-shared-content-client-v1.8.19) (2026-03-21)
 
 ### Bug Fixes